third_party/blink/web_tests/platform/generic/http/tests/security/cross-frame-access-parent-isolated-world-expected.txt

Tests that isolated worlds do not allow cross-origin frame access

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Parent document access from isolated world is not allowed.
PASS successfullyParsed is true

TEST COMPLETE
